fix: address PR review comments
- Add aria-label to chevron CollapsibleTrigger for screen reader accessibility
- Invalidate summary query in contract update/delete mutations to prevent stale totals
- Extract getRenewalRowClass to shared utility in lib/utils.ts, used by both dashboard and upcoming renewals page
